refactor: Centralize problem utilities and CMOP base class, and add tensor dimension utility.
