+++ src/evomo/utils/__init__.py
@@ -1,7 +1,10 @@
 __all__ = [
     "unique_rows_sorted",
     "parse_evaluate",
+    "at_least_2d",
+    "get_pareto_front",
+    "load_pareto_front_from_file",
 ]
 
-from .tensor_ops import unique_rows_sorted
-from .common import parse_evaluate
+from .tensor_ops import unique_rows_sorted, at_least_2d, get_pareto_front
+from .common import parse_evaluate, load_pareto_front_from_file
